modified conversion

+++ gui/game_window.hpp
@@ -40,7 +40,8 @@ public:
     void create_lines();
 
     void read_notes(vector<string> list);
-    void actualize_notes(int spent_time);
+
+    void actualize_notes(float spent_time);
     int getPositionNote(int index);
     int getDurationNote(int index);
     int get_musicline_radius();
